+++ bin/build_vimr.sh
@@ -52,7 +52,13 @@ echo "### Xcodebuilding"
 rm -rf ${build_path}
 
 if [[ ${code_sign} == true ]] ; then
-    xcodebuild CODE_SIGN_IDENTITY="Developer ID Application: Tae Won Ha (H96Q2NKTQH)" -configuration Release -scheme VimR -workspace VimR.xcworkspace -derivedDataPath ${build_path} clean build
+    identity="Developer ID Application: Tae Won Ha (H96Q2NKTQH)"
+    xcodebuild CODE_SIGN_IDENTITY="${identity}" -configuration Release -scheme VimR -workspace VimR.xcworkspace -derivedDataPath ${build_path} clean build
+    pushd ${build_path}/Build/Products/Release > /dev/null
+        codesign --force -s "${identity}" --deep --timestamp --options=runtime VimR.app/Contents/Frameworks/Sparkle.framework/Versions/A/Resources/Autoupdate.app
+        codesign --force -s "${identity}" --deep --timestamp --options=runtime VimR.app/Contents/Frameworks/Sparkle.framework/Versions/A
+        codesign --force -s "${identity}" --deep --timestamp --options=runtime VimR.app
+    popd > /dev/null
 else
     xcodebuild -configuration Release -scheme VimR -workspace VimR.xcworkspace -derivedDataPath ${build_path} clean build
 fi